requirements.txt
pyautogui
keyboard
opencv-python
numpy
pytesseract
ipykernel
pygetwindow
pytest
pillow
tqdm
python-dotenv